Avoid updating the CSS if nothing has changed
To avoid flashing between styles, the code adds the new CSS first, then removes the old one. By adding a size option, some styles might be updated without any changes to the CSS. For example all sizes of the "background" style are identical. In this case, we would add the same CSS twice, then remove it, leaving the compose window unstyled.

To avoid this, we compare the new CSS code with the one already applied (which we already keep to be able remove it), and if nothing has changed, we leave the window untouched.